Add "bui-core.el"
This commit fixes a problem with 'bui-define-interface' macro: the
expanded code (generated by this macro) requires 'bui-BUFFER-TYPE'
feature but it is actually too late for this, because the macro needs
'bui-BUFFER-TYPE-symbol-specifications' variable at expansion time.

This could be fixed by requiring both 'bui-BUFFER-TYPE' features by
'bui' itself, however since 'bui-list' and 'bui-info' require 'bui' in
turn, the most part of 'bui' is moved to the new 'bui-core' feature.

* bui.el: Move everything except 'bui-define-entry-type' and
'bui-define-interface' macros to...
* bui-core.el: ... here.  New file.
* bui-info.el: Require it instead of 'bui'.
* bui-list.el: Likewise.